refactor: Util에 있는 Validation을 Validator class로 분리
