+++ src/main/java/baseball/util/IntegerUtil.java
@@ -2,20 +2,12 @@ package baseball.util;
 
 public class IntegerUtil {
 
-    private static final String INVALID_INTEGER_MESSAGE = "이 값은 정수로 변환할 수 없습니다.";
-
     public static boolean isInteger(String string) {
         try {
             Integer.parseInt(string);
             return true;
         } catch (NumberFormatException ex) {
             return false;
-        }
-    }
-
-    public static void validateInteger(String string) {
-        if (!isInteger(string)) {
-            ExceptionUtil.throwInvalidValueException(INVALID_INTEGER_MESSAGE);
         }
     }
